caldav_util.c: move caldav_create_defaultcalendars() from http_caldav.c and parameterize it
